pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shadowbyte"
version = "2.0.0"
description = "CLI Toolkit"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Gerijacki", email = "gerijackidev@gmail.com" },
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "yt-dlp>=2023.0.0",
    "requests>=2.31.0",
    "qrcode[pil]>=7.0.0",
    "cryptography>=41.0.0",
    "scapy>=2.5.0",
    "speedtest-cli>=2.1.3",
    "py-cpuinfo>=9.0.0",
]

[project.scripts]
shadowbyte = "shadowbyte.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/shadowbyte"]

[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
